+++ projects/ontimize-web-ngx/src/lib/components/o-service-component.class.ts
@@ -293,6 +293,9 @@ export abstract class AbstractOServiceComponent<T extends AbstractComponentState
     if (route.length > 0) {
       const relativeTo = this.recursiveInsert ? this.actRoute.parent : this.actRoute;
       const qParams = {};
+      if (this.formLayoutManager && this.formLayoutManager.isTabMode()) {
+        qParams[Codes.INSERTION_MODE] = 'true';
+      }
       this.navigateToDetail(route, qParams, relativeTo);
     }
   }